refactor: extract the editor metabox rendering into its own class
Moves 57 methods and ~1080 lines out of Documentate_Documents into
Documentate_Document_Meta_Boxes, following the shape wp-decker uses for its
event CPT: the post type keeps the hooks, a collaborator does the work.

The boundary is exact. After the previous commit moved the shared field-value
helpers to Documents_Meta_Handler, the rendering block reaches nothing that
stays behind - no methods, no properties, no constants. That was the point of
doing the helpers first; before it there were five calls crossing the line.

register_meta_boxes(), render_type_metabox() and render_sections_metabox()
stay on the CPT as delegators. The first is a hook callback, and the other two
are called directly by 59 tests, so all three keep their published home while
their bodies move.

Three imports in the CPT are now unused and dropped. One of them,
SchemaConverter, was already unused before this commit.

About 60 tests reflected on private methods that moved. The two files whose
every target moved now instantiate the renderer directly; the two mixed files
resolve the owner, which for DocumentateDocumentsHelpersTest means its
invoke() helper picks whichever object still declares the method.

Markup verified identical, sections and type metabox both, by rendering a
schema covering every control type against the pre-extraction code.

The CPT is now 1717 lines from 3357, its complexity 245 from 446, and
ExcessiveClassLength is resolved. Plugin-wide the alert count does not move:
the new class carries an ExcessiveClassComplexity of 204 of its own. Splitting
the class was never going to fix that by itself - what it does is make the
remaining two extractions possible, after which the rendering class can be
divided along the scalar/repeater line it already has.
