+++ tests/unit/includes/custom-post-types/DocumentateSchemaRowTest.php
@@ -13,7 +13,7 @@ class DocumentateSchemaRowTest extends WP_UnitTestCase {
 	/**
 	 * Instance under test.
 	 *
-	 * @var Documentate_Documents
+	 * @var Documentate_Document_Meta_Boxes
 	 */
 	private $documents;
 
@@ -23,7 +23,7 @@ class DocumentateSchemaRowTest extends WP_UnitTestCase {
 	public function set_up() {
 		parent::set_up();
 
-		$this->documents = new Documentate_Documents();
+		$this->documents = new Documentate_Document_Meta_Boxes();
 	}
 
 	/**
